-Integrate Jacob's changes -(DONE)Add background to fragments and buttons to not allow overlap to distract -Disallow Uncategorized from being deleted -(DONE)Figure out langauge sortin -Add abbreviations to phrases on both edit AND main activity screens -(DONE)Add a way to delete a language -Add a back button to edit screen
Noticed bugs
When a phrase is deleted and then the add phrase fragment is brought up, that phrase is still there.
Similarly, newly added phrases dont show up if the add phrase fragment is brought up more than once within a single session of EditActivity
(FIXED)The phrases are also getting added to uncategorized
If an already existing phrase is having language overwritten, the save button does nothing and we're stuck on the "New Recording" Screen
(FIXED)Adding a new langauge to a phrase that already exists adds a duplicate phrase to the list.

Some possible enhancements:
Add a label above the spinners in the add phrase fragment
For purposes of avoiding substring errors, make '[' and ']' illegal characters in language and language abbreviation